move block_number to model.context
