+++ credmark/cmf/types/__init__.py
@@ -2,7 +2,6 @@
 
 from .account import Account, Accounts
 from .address import Address, NULL_ADDRESS
-from .block_number import BlockNumber
 from .contract import Contract, Contracts
 from .ledger import ContractLedger
 from .portfolio import Portfolio
